extrace encode from braille method into a decoder module
